backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.17
pydantic==2.5.0
python-multipart==0.0.6
better-exceptions==0.3.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography>=41.0.7
python-dotenv==1.0.0
alembic==1.13.1
asyncpg==0.29.0
pytz==2023.3
uuid==1.30
cohere==5.5.3